backend/requirements.txt

fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
tensorflow
numpy
pillow
python-dotenv
requests
aiofiles
pymysql
